+++ packages/google_maps_flutter/google_maps_flutter_android/lib/src/google_maps_flutter_android.dart
@@ -545,12 +545,8 @@ class GoogleMapsFlutterAndroid extends GoogleMapsFlutterPlatform {
     };
   }
 
-  /// Asks the Google Maps SDK to do the thread-blocking work it normally does
-  /// when a map is shown for the first time.
-  ///
-  /// This gives the developer the option to move that jank to a different
-  /// part of the map (typically, the app startup, where missed frames
-  /// aren't going to be noticed).
+  /// Attempts to trigger any thread-blocking work
+  /// the Google Maps SDK normally does when a map is shown for the first time.
   Future<void> warmup() async {
     await _initializerApi.warmup();
   }